Removed redundant casting

+++ test/base-conversions.js
@@ -4,7 +4,7 @@ function base26ToBase10(base26) {
         let x = 25 - i;
         let digit = BigInt(shifts.indexOf(base26[x]));
         let value = BigInt(26) ** BigInt(i);
-        base10 = BigInt(base10 + digit * value);
+        base10 = base10 + digit * value;
     };
     return base10;
 };
@@ -20,11 +20,11 @@ function base10ToBase26(base10) {
     return base26
 };
 
-const shifts = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K','L', 'M', 'N', 'O', 'P'];
+const shifts = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P'];
 var base26 = "BB3F6L2K5N942KF944HOAL15O0";
 console.log(base26ToBase10(base26));  // BB3F6L2K5N942KF944HOAL15O0 -> 2705942045828863551735381465367244636n
 
-var base10 = BigInt("6156119580207157310796674288400203775");
+var base10 = BigInt("6156119580207157310796674288400203775");  // Must be type BigInt
 console.log(base10ToBase26(base10));  // 6156119580207157310796674288400203775 -> PPPPPPPPPPPPPPPPPPPPPPPPPP
 
 console.log(base10ToBase26(base26ToBase10(base26)));  // BB3F6L2K5N942KF944HOAL15O0 -> BB3F6L2K5N942KF944HOAL15O0
